+++ routes/places.routes.js
@@ -1,8 +1,10 @@
-const Place = require("../models/Place.model");
-const router = require("express").Router();
+const isAdmin = require('../middleware/isAdmin');
+const isLoggedIn = require('../middleware/isLoggedin');
+const Place = require('../models/Place.model');
+const router = require('express').Router();
 
 //Fetch all places
-router.get("/", async (req, res, next) => {
+router.get('/', async (req, res, next) => {
   try {
     const foundPlaces = await Place.find();
     res.status(200).json(foundPlaces);
@@ -12,7 +14,7 @@ router.get("/", async (req, res, next) => {
 });
 
 //Fetch place by id
-router.get("/:id", async (req, res, next) => {
+router.get('/:id', async (req, res, next) => {
   try {
     const foundPlace = await Place.findById(req.params.id);
     res.status(200).json(foundPlace);
@@ -21,25 +23,41 @@ router.get("/:id", async (req, res, next) => {
   }
 });
 
+//Create place
+router.post('/', isLoggedIn, async (req, res, next) => {
+  try {
+    const createdPlace = await Place.create(req.body);
+    res.status(201).json(createdPlace);
+  } catch (error) {
+    res.sendStatus(400);
+  }
+});
+
 //Update place
-router.post("/:id", async (req, res, next) => {
+router.patch('/:id', isLoggedIn, async (req, res, next) => {
   try {
     const id = req.params.id;
-    const updatedPlace = await Place.findByIdAndUpdate(id, req.body, {
-      new: true,
-    });
-    res.status(200).json(updatedPlace);
+    const foundPlace = await Place.findById(id);
+    if (foundPlace.user.toString() === req.user._id.toString() || isAdmin) {
+      const updatedPlace = await Place.findByIdAndUpdate(id, req.body, {
+        new: true,
+      });
+      res.status(200).json(updatedPlace);
+    }
   } catch (error) {
     res.sendStatus(400);
   }
 });
 
 //Delete place
-router.delete("/:id", async (req, res, next) => {
+router.delete('/:id', isLoggedIn, async (req, res, next) => {
   try {
     const id = req.params.id;
-    const deletedPlace = await Place.findByIdAndDelete(id);
-    res.status(200).json(deletedPlace);
+    const foundPlace = await Place.findById(id);
+    if (foundPlace.user.toString() === req.user._id.toString() || isAdmin) {
+      const deletedPlace = await Place.findByIdAndDelete(id);
+      res.status(200).json(deletedPlace);
+    }
   } catch (error) {
     res.sendStatus(404);
   }
